feat: SCAN-1611 future nft support
