+++ src/app/components/NFTPreview/NFTInfo.ts
@@ -75,16 +75,16 @@ export const NFTNames = {
       zh: '嘻哈烤仔',
     },
   },
-  ancientChineseGod: {
-    'Pan Gu': '盘古',
-    'Nü Wa': '女娲',
-    'Fu Xi': '伏羲',
-    'Sui Ren': '燧人',
-    'Shen Nong': '神农',
-    Yu: '禹',
-    'Yellow Emperor': '黄帝',
-    'Chi You': '蚩尤',
-    'Gong Gong': '共工',
-    'Kua Fu': '夸父',
-  },
+  // ancientChineseGod: {
+  //   'Pan Gu': '盘古',
+  //   'Nü Wa': '女娲',
+  //   'Fu Xi': '伏羲',
+  //   'Sui Ren': '燧人',
+  //   'Shen Nong': '神农',
+  //   Yu: '禹',
+  //   'Yellow Emperor': '黄帝',
+  //   'Chi You': '蚩尤',
+  //   'Gong Gong': '共工',
+  //   'Kua Fu': '夸父',
+  // },
 };
